refactor: reorganize patterns in its own file
